test: fold oauth-connect card scope tests into choice-copy-surfaces suite (#40111)
