+++ clients/web/src/domains/chat/components/surfaces/choice-copy-surfaces.test.tsx
@@ -32,6 +32,7 @@ import { OAuthConnectSurface } from "@/domains/chat/components/surfaces/oauth-co
 import { SurfaceRouter } from "@/domains/chat/components/surfaces/surface-router";
 import type {
   ManagedOAuthConnectClient,
+  ManagedOAuthConnectOptions,
   ManagedOAuthConnectResult,
 } from "@/domains/chat/api/managed-oauth";
 import { assistantsOauthConnectionsListQueryKey } from "@/generated/api/@tanstack/react-query.gen";
@@ -67,6 +68,23 @@ function makeSurface(overrides: Partial<Surface>): Surface {
     surfaceType: "choice",
     data: {},
     ...overrides,
+  };
+}
+
+function makeConnectedResult(
+  scopesGranted: string[],
+): ManagedOAuthConnectResult {
+  return {
+    status: "connected",
+    connection: {
+      id: "conn-1",
+      provider: "google",
+      status: "ACTIVE",
+      connected: true,
+      account_label: "user@example.com",
+      scopes_granted: scopesGranted,
+      expires_at: null,
+    } as OAuthConnection,
   };
 }
 
@@ -222,18 +240,7 @@ describe("OAuthConnectSurface", () => {
     const onAction = mock(() => {});
     const oauthClient: ManagedOAuthConnectClient = {
       fetchProvider: mock(async () => null),
-      connect: mock(async () => ({
-        status: "connected" as const,
-        connection: {
-          id: "conn-1",
-          provider: "google",
-          status: "ACTIVE",
-          connected: true,
-          account_label: "user@example.com",
-          scopes_granted: ["gmail.readonly"],
-          expires_at: null,
-        } as OAuthConnection,
-      })),
+      connect: mock(async () => makeConnectedResult(["gmail.readonly"])),
     };
 
     const { getByRole, queryByText, invalidateQueries } = renderWithQueryClient(
@@ -292,6 +299,75 @@ describe("OAuthConnectSurface", () => {
     });
   });
 
+  test("omits requestedScopes when the surface data carries none", async () => {
+    const connect = mock(async (_options: ManagedOAuthConnectOptions) =>
+      makeConnectedResult([]),
+    );
+    const oauthClient: ManagedOAuthConnectClient = {
+      fetchProvider: mock(async () => null),
+      connect,
+    };
+
+    const { getByRole } = renderWithQueryClient(
+      <OAuthConnectSurface
+        surface={makeSurface({
+          surfaceType: "oauth_connect",
+          data: { providerKey: "google", displayName: "Google" },
+        })}
+        assistantId="assistant-1"
+        oauthClient={oauthClient}
+        onAction={mock(() => {})}
+      />,
+    );
+
+    fireEvent.click(getByRole("button", { name: "Connect" }));
+
+    await waitFor(() => expect(connect).toHaveBeenCalledTimes(1));
+    expect(connect.mock.calls[0]?.[0]?.requestedScopes).toBeUndefined();
+  });
+
+  test("reports scopesGranted from the resulting connection, not the request", async () => {
+    // The platform decides what was actually granted; the action payload must
+    // reflect the connection's scopes_granted so the model can verify the
+    // grant includes what it asked for.
+    const onAction = mock(() => {});
+    const oauthClient: ManagedOAuthConnectClient = {
+      fetchProvider: mock(async () => null),
+      connect: mock(async () =>
+        makeConnectedResult(["gmail.readonly", "tasks", "calendar"]),
+      ),
+    };
+
+    const { getByRole } = renderWithQueryClient(
+      <OAuthConnectSurface
+        surface={makeSurface({
+          surfaceType: "oauth_connect",
+          data: {
+            providerKey: "google",
+            displayName: "Google",
+            requestedScopes: ["tasks"],
+          },
+        })}
+        assistantId="assistant-1"
+        oauthClient={oauthClient}
+        onAction={onAction}
+      />,
+    );
+
+    fireEvent.click(getByRole("button", { name: "Connect" }));
+
+    await waitFor(() => {
+      expect(onAction).toHaveBeenCalledWith("surface-1", "connect", {
+        status: "connected",
+        providerKey: "google",
+        providerLabel: "Google",
+        connectionId: "conn-1",
+        accountLabel: "user@example.com",
+        scopesGranted: ["gmail.readonly", "tasks", "calendar"],
+      });
+    });
+  });
+
   test("does not submit the surface action after the card unmounts mid-connect", async () => {
     // A remounted card can await the same deduped OAuth promise; only the
     // still-mounted instance may report the result, so one authorization
@@ -326,18 +402,7 @@ describe("OAuthConnectSurface", () => {
 
     // The transcript re-render replaced this instance while OAuth was in flight.
     unmount();
-    resolveConnect({
-      status: "connected",
-      connection: {
-        id: "conn-1",
-        provider: "google",
-        status: "ACTIVE",
-        connected: true,
-        account_label: "user@example.com",
-        scopes_granted: [],
-        expires_at: null,
-      } as OAuthConnection,
-    });
+    resolveConnect(makeConnectedResult([]));
     await new Promise((resolve) => setTimeout(resolve, 0));
 
     expect(onAction).not.toHaveBeenCalled();
